move prisma lib
