+++ libs/prisma/src/index.ts
@@ -1,0 +1,2 @@
+export * from './prisma.module';
+export * from './prisma.service';
