+++ package.json
@@ -99,7 +99,8 @@
     "moduleNameMapper": {
       "^src/(.*)$": "<rootDir>/src/$1",
       "^test/(.*)$": "<rootDir>/test/$1",
-      "^@lib/logger(|/.*)$": "<rootDir>/libs/logger/src/$1"
+      "^@lib/logger(|/.*)$": "<rootDir>/libs/logger/src/$1",
+      "^@lib/prisma(|/.*)$": "<rootDir>/libs/prisma/src/$1"
     },
     "coverageDirectory": "../coverage",
     "testEnvironment": "node",
