+++ src/client/client.repository.ts
@@ -1,4 +1,5 @@
 import { Loggable } from '@lib/logger/decorator/loggable';
+import { PrismaService } from '@lib/prisma';
 import {
   ConflictException,
   ForbiddenException,
@@ -8,7 +9,6 @@ import {
 } from '@nestjs/common';
 import { Client } from '@prisma/client';
 import { PrismaClientKnownRequestError } from '@prisma/client/runtime/library';
-import { PrismaService } from 'src/prisma/prisma.service';
 
 @Injectable()
 @Loggable()
