+++ src/role/role.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { RoleController } from './role.controller';
 import { RoleService } from './role.service';
-import { PrismaModule } from 'src/prisma/prisma.module';
 import { RoleRepository } from './role.repository';
 import { AuthModule } from 'src/auth/auth.module';
 import { LoggerModule } from '@lib/logger';
+import { PrismaModule } from '@lib/prisma';
 
 @Module({
   imports: [PrismaModule, AuthModule, LoggerModule],
