+++ src/user/user.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { UserService } from './user.service';
-import { PrismaModule } from 'src/prisma/prisma.module';
 import { UserRepository } from './user.repository';
 import { LoggerModule } from '@lib/logger';
+import { PrismaModule } from '@lib/prisma';
 
 @Module({
   imports: [PrismaModule, LoggerModule],
