+++ test/jest-e2e.json
@@ -13,6 +13,8 @@
   "moduleNameMapper": {
     "^src/(.*)$": "<rootDir>/../src/$1",
     "@lib/logger/(.*)": "<rootDir>/../libs/logger/src/$1",
-    "@lib/logger": "<rootDir>/../libs/logger/src"
+    "@lib/logger": "<rootDir>/../libs/logger/src",
+    "@lib/prisma/(.*)": "<rootDir>/../libs/prisma/src/$1",
+    "@lib/prisma": "<rootDir>/../libs/prisma/src"
   }
 }
