+++ tsconfig.json
@@ -23,6 +23,12 @@
       ],
       "@lib/logger/*": [
         "libs/logger/src/*"
+      ],
+      "@lib/prisma": [
+        "libs/prisma/src"
+      ],
+      "@lib/prisma/*": [
+        "libs/prisma/src/*"
       ]
     }
   }
